+++ src/main/java/com/chat/websocket/service/impl/ConversationImpl.java
@@ -8,6 +8,7 @@ import com.chat.websocket.repository.GroupMemberRepository;
 import com.chat.websocket.service.ConversationService;
 import com.chat.websocket.service.GroupMemberService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -24,22 +25,17 @@ public class ConversationImpl implements ConversationService {
     @Autowired
     private GroupMemberRepository groupMemberRepository;
 
-    private GroupMemberService groupMemberService;
     @Autowired
-    public GroupMemberService getGroupMemberService() {
-        return groupMemberService;
-    }
-    public void setGroupMemberService(GroupMemberService groupMemberService) {
-        this.groupMemberService = groupMemberService;
-    }
+    @Lazy
+    private GroupMemberService groupMemberService;
 
     @Override
-    public List<Conversation> findAll() {
+    public List<Conversation> getAllConversation() {
         return conversationRepository.findAll();
     }
 
     @Override
-    public Conversation findByID(int id) {
+    public Conversation getConversationByID(int id) {
         Optional<Conversation> conversationResult = conversationRepository.findById(id);
         Conversation conversation = conversationResult.get();
 
@@ -47,52 +43,68 @@ public class ConversationImpl implements ConversationService {
     }
 
     @Override
-    public Conversation saveReturnSaved(Conversation conversation) {
+    public Conversation addConversationAndReturnConversationSaved(Conversation conversation) {
         Conversation conversationSaved = conversationRepository.save(conversation);
         return conversationSaved;
     }
 
     @Override
-    public void save(Conversation conversation) {
+    public void addConversation(Conversation conversation) {
         conversationRepository.save(conversation);
     }
 
     @Override
-    public void deleteByID(int id) {
-        conversationRepository.deleteById(id);
+    public Conversation deleteConversationByID(int id) {
+        Optional<Conversation> conversationResult = conversationRepository.findById(id);
+
+        if (conversationResult.isPresent()) {
+            conversationRepository.deleteById(id);
+            return conversationResult.get();
+        }
+
+        return null;
     }
 
     @Override
     public List<Conversation> findConversationsByContactID(int contactID) {
         // Conversation list
-        List<Conversation>  conversations = new ArrayList<>();
+        List<Conversation> conversations = new ArrayList<>();
 
         // Get GroupMembers by ContactID
         List<GroupMember> groupMembers = groupMemberService.findGroupMembersByContactID(contactID);
 
-        for (GroupMember tempGroupMember:
-             groupMembers) {
+        if (groupMembers != null ) {
+            for (GroupMember tempGroupMember:
+                    groupMembers) {
 
-            // Get Conversation by GroupMember
-            Conversation conversation = findConversationByGroupMemberID(tempGroupMember.getGroupMemberID());
+                // Get Conversation by GroupMember
+                Conversation conversation = findConversationByGroupMemberID(tempGroupMember.getGroupMemberID());
 
-            conversations.add(conversation);
+                conversations.add(conversation);
+            }
+
+            return conversations;
         }
 
-        return conversations;
+        return null;
     }
 
     @Override
     public Conversation findConversationByGroupMemberID(int groupMemberID) {
         // Group member
         Optional<GroupMember> groupMemberResult = groupMemberRepository.findById(groupMemberID);
-        GroupMember groupMember = groupMemberResult.get();
 
-        // find Conversation by group member is
-        int conversationID = groupMember.getConversation().getConversationID();
-        Conversation conversation = findByID(conversationID);
+        if (groupMemberResult.isPresent()) {
+            GroupMember groupMember = groupMemberResult.get();
 
-        return conversation;
+            // find Conversation by group member is
+            int conversationID = groupMember.getConversation().getConversationID();
+            Conversation conversation = getConversationByID(conversationID);
+
+            return conversation;
+        }
+
+        return null;
     }
 
 }
